+++ src/schemas/api/index.ts
@@ -1,5 +1,5 @@
 export * from './error_v0'
+export * from './exports_v0'
 export * from './jobs_v0'
 export * from './me_v0'
 export * from './products_v0'
-export * from './reports_v0'