+++ src/schemas/database/project/expandedProject_v0.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { APIReportSchema_v0 } from '../../api/reports_v0'
+import { APIExportSchema_v0 } from '../../api/exports_v0'
 import {
     disableUserWrite,
     getPatchSchema,
@@ -34,8 +34,8 @@ export const ProjectSubcollectionsSchema_v0 = z.object({
     graph: GraphSchema_v0.optional(),
     pipeTypes: z.record(z.string(), PipeTypeDataSchema_v0).optional(),
     registerTypes: z.record(z.string(), RegisterTypeDataSchema_v0).optional(),
-    reports: disableUserWrite(
-        z.record(z.string(), APIReportSchema_v0).optional()
+    exports: disableUserWrite(
+        z.record(z.string(), APIExportSchema_v0).optional()
     ),
     roofTypes: z.record(z.string(), RoofTypeDataSchema_v0).optional(),
     sheetFiles: z.record(z.string(), SheetFileDataSchema_v0).optional(),